requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
pymongo==4.5.0
celery==5.3.1
redis==4.6.0
twilio==8.5.0
requests==2.31.0
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
Werkzeug==2.3.7
gunicorn==21.2.0
APScheduler==3.10.4
boto3==1.34.0
botocore==1.34.0
Pillow==10.0.1
qrcode[pil]==7.4.2
pymongo[srv]==4.5.0
razorpay==1.3.0
stripe==5.5.0
redis==4.6.0
email-validator==2.0.0
PyJWT==2.8.0 